Refactor parsing code out of main
Refactor parsing code out of main to its own package
